src/llm/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "This is the container for the machine learning pipeline portion of the project"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.1.1",
    "einops>=0.8.1",
    "fastapi>=0.118.0",
    "google-cloud-storage>=2.19.0",
    "langchain-google-vertexai>=2.1.2",
    "numpy>=2.3.3",
    "pydantic>=2.12.0",
    "python-dotenv>=1.1.1",
    "rank-bm25>=0.2.2",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.1",
    "uvicorn>=0.37.0",
    "datasets>=2.18.0",
    "pyarrow>=15.0.0",
    "google-cloud-storage>=2.10.0",
    "google-cloud-aiplatform>=1.70.0",
]
